Update Android and JS implementations to Tabris.js 3.4
Also react on onNewIntent events. Previously, "urlLaunch" was only
triggered when the activity was started initially.

Replace cordova-plugin-customurlscheme used in the example with own URL
scheme configuration. The plugin was previously only used for
configuration purposes.

Demonstrate delegation of onNewIntent events by using the singleTask
launch mode. When using `app.launch()`, it will enforce the activity to
be launched at the root of a separate task instead of as part of the
launcher's task and an existing activity to be brought to the foreground
instead of launching a new activity.

Move example to .example to enable builds that were previously not
possible due to a Cordova limitation [1]. Note that the example build
only works with Tabris CLI <= 3.2 since it does not work with the newer
Cordova versions supplied with Tabris CLI 3.3+. Long-term solution
pending.

[1]: https://github.com/apache/cordova-cli/issues/401

Change-Id: I3691181fdd5d68157ddee2efa53b651d7a05a4c4