Ensure `urlLaunch` event is triggered correctly
Previously, the `urlLaunch` event handler was not being called when the
app was launched through a URL, and the app hadn't been opened at all.
The `urlLaunch` event was fired before the corresponding JavaScript
handler on the app's side was initiated.

Run the JavaScript handler invoke within a new Runnable and delay the
execution start by 0 milliseconds.

Replace the standalone instance of the `Handler()` class with a scope
handler.